# Optimize python QgsFeature.setAttribute(s) calls