--- a/python/core/auto_generated/qgsattributes.sip.in
+++ b/python/core/auto_generated/qgsattributes.sip.in
@@ -81,6 +81,22 @@ typedef QVector<QVariant> QgsAttributes;
     {
       qv->append( QVariant( QVariant::Int ) );
     }
+    else if ( PyBool_Check( obj ) )
+    {
+      qv->append( QVariant( PyObject_IsTrue( obj ) == 1 ) );
+    }
+    else if ( PyLong_Check( obj ) )
+    {
+      qv->append( QVariant( PyLong_AsLongLong( obj ) ) );
+    }
+    else if ( PyFloat_Check( obj ) )
+    {
+      qv->append( QVariant( PyFloat_AsDouble( obj ) ) );
+    }
+    else if ( PyUnicode_Check( obj ) )
+    {
+      qv->append( QVariant( QString::fromUtf8( PyUnicode_AsUTF8( obj ) ) ) );
+    }
     else
     {
       int state;