@nyalldawson 2024-05-17
FYI the kdab hotspot tool is super useful in identifying areas for optimisation like this