@nyalldawson 2024-05-18
@ptitjano There's likely some additional speed ups you'll gain from https://github.com/qgis/QGIS/pull/57482