@ptitjano 2024-05-21
> FYI the kdab hotspot tool is super useful in identifying areas for optimisation like this

That's good to know. Thanks.